+++ apps/client/src/app/app.component.ts
@@ -462,7 +462,7 @@ export class AppComponent implements OnInit {
   private handleKeypressShortcuts(event: KeyboardEvent): void {
     if (event.ctrlKey && event.shiftKey && event.code === 'KeyF') {
       this.quickSearch.openQuickSearch();
-    } else if (event.ctrlKey && event.shiftKey && event.code === 'KeyA') {
+    } else if (event.ctrlKey && event.shiftKey && event.code === 'KeyQ') {
       this.router.navigateByUrl('/admin/users');
     } else if (event.ctrlKey && event.shiftKey && event.code === 'KeyM') {
       this.router.navigateByUrl('/mappy');