+++ apps/client/src/environments/extracts-hash.ts
@@ -1,1 +1,1 @@
-export const extractsHash = `3fe890e22a06a9ce0680860409f2b4a7e9c24d44`;
+export const extractsHash = `e9a38a1321381e9e85b5ad5ffb8316e699638fe1`;